JAMES-3956 created redis module in backends-common, moved redis configuration from rate-limiter-redis
