+++ code-coverage-report/pom.xml
@@ -344,6 +344,11 @@
             <artifactId>apache-james-backends-rabbitmq</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
+            <artifactId>apache-james-backends-redis</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!--        7. james-core-->
         <dependency>
